[refactor] placeService 분리
